Add forward declarations for local functions.
When functions are called before they are defined their name is later
searched in global scope upon execution (and the result often found to
be nil). So either we maintain a proper order of functions definitions
within our lua file or we just add forward declarations at the top of
the file. With this commit we've done the latter.